feat: add error validation about invalid mentions on bsky
